created the counter, Theming, and other changes
	modified:   README.md
	modified:   counter.php
	modified:   counters/index.php
	new file:   generator.php
	modified:   index.php
	renamed:    templates/custom.html -> templates/custom.tpl
	deleted:    templates/pociot.html
	new file:   templates/pociot.tpl
	deleted:    templates/span.html
	new file:   themes.php

+++ counter.php
@@ -1,6 +1,11 @@
 <?php
 declare(strict_types=1);
 
+if (!isset($_SESSION))
+{
+    session_start();
+}
+
 //Contador de visitas simples
 //A basic visitor count
 
@@ -13,11 +18,13 @@ function getCounterName(string $input)
     return $nome;
 }
 
-$contador = getCounterName($_GET['site'] ?? $_GET['link'] ?? '');
+$url = $_GET['url'] ?? '';
 
-$counter_title   = "Visitor";
+$counter_name = filter_var($url, FILTER_VALIDATE_URL) ? getCounterName($url) : getCounterName('');
+
+$counter_label   = "Visitors";
 $counter_dir     = __DIR__."/counters/";
-$counter_file    = "{$counter_dir}{$contador}.ctr";
+$counter_file    = "{$counter_dir}{$counter_name}.ctr";
 
 if(!file_exists($counter_dir))
 {
@@ -29,51 +36,75 @@ if(!file_exists($counter_file))
     file_put_contents($counter_file, '');
 }
 
-$counter_title = $_GET['texto'] ?? $counter_title;
+$counter_label = $_GET['label'] ?? $counter_label;
 
 if (($_GET['clear_count'] ?? null) == 'CLEAR')
 {
     file_put_contents($counter_file, 0);
 }else{
     $count = (int) file_get_contents($counter_file);
-    $count = $count + 1;
+
+    $last_count = $_SESSION['last_count'] ?? null;
+
+    $_SESSION['last_count'] = $last_count ?? [
+        'count' => $count,
+        'time'  => time(),
+    ];
+
+    if(($last_count['time'] ?? null) < (time() - 3))
+    {
+        $count++;
+        file_put_contents($counter_file, $count);
+        $_SESSION['last_count'] = [
+            'count' => $count,
+            'time'  => time(),
+        ];
+    }
+
     file_put_contents($counter_file, $count);
 }
 
-if (is_numeric($_GET['set_count'] ?? null))
+if ( !file_exists($counter_file) && is_numeric($_GET['set_count'] ?? null))
 {
     $set_count  = (int) $_GET['set_count'];
     $count      = $set_count > 0 ? $set_count - 1 : 0;
     file_put_contents($counter_file, $count);
 }
 
-$template = $_GET['template'] ?? 'pociot';
-$template_content = file_exists(__DIR__."/templates/{$template}.html")
-    ? file_get_contents(__DIR__."/templates/{$template}.html")
-    : file_get_contents(__DIR__."/templates/span.html");
+$themes = require __DIR__.'/themes.php';
 
-$radius     = is_numeric($radius = $_GET['radius'] ?? '5') ? (int) $radius : 5;
-$color_1    = $_GET['color_1'] ?? '#00ff00';
-$color_2    = $_GET['color_2'] ?? $color_1;
-$color_1_opacity = $_GET['color_1_opacity'] ?? '0.5';
-$color_2_opacity = $_GET['color_2_opacity'] ?? $color_1_opacity;
-$text_color  = $_GET['text_color'] ?? '#000';
-$title_color = $_GET['title_color'] ?? $text_color;
-$count_color = $_GET['count_color'] ?? $title_color;
+$template = $_GET['template'] ?? 'pociot';
+
+$template_content = file_exists(__DIR__."/templates/{$template}.tpl")
+    ? file_get_contents(__DIR__."/templates/{$template}.tpl")
+    : file_get_contents(__DIR__."/templates/pociot.tpl");
+
+$theme = $themes[$template] ?? $themes['custom'];
+$radius             = is_numeric($radius = $_GET['radius'] ?? '5') ? (int) $radius : 5;
+$color_1            = $_GET['color_1']         ?? $theme['color_1'];
+$color_2            = $_GET['color_2']         ?? $theme['color_2'];
+$color_1_opacity    = $_GET['color_1_opacity'] ?? $theme['color_1_opacity'];
+$color_2_opacity    = $_GET['color_2_opacity'] ?? $theme['color_2_opacity'];
+$text_color         = $_GET['text_color']      ?? $theme['text_color'];
+$label_color        = $_GET['label_color']     ?? $theme['label_color'];
+$count_color        = $_GET['count_color']     ?? $theme['count_color'];
+
+// die(var_dump($_GET));
 
 if($count ?? null)
 {
     $replaces = [
-        '##TITLE##' => $counter_title,
-        '##COUNT##' => $count,
-        '##RADIUS##' => $radius,
-        '##COLOR_1##' => $color_1,
-        '##COLOR_2##' => $color_2,
-        '##COLOR_1_OPACITY##' => $color_1_opacity,
-        '##COLOR_2_OPACITY##' => $color_2_opacity,
-        '##TITLE_COLOR##' => $title_color,
-        '##COUNT_COLOR##' => $count_color,
+        '##LABEL##'             => $counter_label,
+        '##COUNT##'             => $count,
+        '##RADIUS##'            => $radius,
+        '##COLOR_1##'           => $color_1,
+        '##COLOR_2##'           => $color_2,
+        '##COLOR_1_OPACITY##'   => $color_1_opacity,
+        '##COLOR_2_OPACITY##'   => $color_2_opacity,
+        '##LABEL_COLOR##'       => $label_color,
+        '##COUNT_COLOR##'       => $count_color,
     ];
-
+    // die(var_dump($replaces));
+    header( 'Content-type: image/svg+xml' );
     echo $template_content = str_replace(array_keys($replaces), array_values($replaces), $template_content);
 }
